renamed rule directory
