+++ src/runner.ts
@@ -2,8 +2,8 @@ import { summary } from "@actions/core";
 import { parse } from "yaml";
 
 import { Inputs } from ".";
-import { ConfigurationFile, Reviewers, Rule } from "./file/types";
-import { validateConfig, validateRegularExpressions } from "./file/validator";
+import { ConfigurationFile, Reviewers, Rule } from "./rules/types";
+import { validateConfig, validateRegularExpressions } from "./rules/validator";
 import { PullRequestApi } from "./github/pullRequest";
 import { TeamApi } from "./github/teams";
 import { ActionLogger, CheckData } from "./github/types";
