+++ src/test/runner/andRule.test.ts
@@ -3,7 +3,7 @@
 /* eslint-disable @typescript-eslint/no-unsafe-call */
 import { mock, MockProxy } from "jest-mock-extended";
 
-import { AndRule } from "../../file/types";
+import { AndRule } from "../../rules/types";
 import { PullRequestApi } from "../../github/pullRequest";
 import { TeamApi } from "../../github/teams";
 import { ActionRunner } from "../../runner";
